+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.1.5]
+
+### Added
+
+- Investigator Creator UI
+
+### Changed
+
+- Replaced Background header with Investigator Creator 
+
 ## [0.1.4]
 
 ### Added